Move parseRequest in to forms
